exercise_3/requirements.txt
flask==1.0.2
python-telegram-bot==11.1.0
requests==2.21.0